tests/requirements.txt
pandas>=1.0.3
rpy2>=3.3.2
GitPython>=3.1.2
nose>=1.3.7
